requirements.txt
customtkinter
openpyxl
docxtpl
docx2pdf
pdf2image
qrcode
Pillow
Faker
python-docx
